adding a geolocate middleware
