+++ packages/api/src/middleware/index.js
@@ -6,3 +6,4 @@ export { default as checkKubernetes } from './check-kubernetes'
 export { default as hardcodedNodes } from './hardcoded-nodes'
 export { default as insecureTest } from './insecure-test'
 export { default as tracking } from './tracking'
+export { default as geolocateMiddleware } from './geolocate'
